CodingSessionService and Input.Parser refactored
